This fix compilation broken in #194

+++ source/sdl/sdl_time.c
@@ -19,5 +19,5 @@ unsigned int Sys_Milliseconds( void )
 uint64_t Sys_Microseconds( void )
 {
 	static Uint64 base = SDL_GetPerformanceCounter();
-	return ( Uint64 )1000000 * ( SDL_GetPerformanceCounter() - base ) / freq;
+	return 1000000ULL * ( SDL_GetPerformanceCounter() - base ) / freq;
 }
